# backend/assets/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "# Nexus Tasks REST API\n\nNexus Tasks is a collaborative project and task management platform.\nThis document describes the full REST API surface of the backend service.\n\n## Authentication\n\nMost endpoints require an authenticated session established via the `/api/v2/auth/login`\nor `/api/v2/auth/register` endpoints. Authentication state is maintained via an\nHTTP-only session cookie (`nexus_user`). Alternatively, machine clients may\nauthenticate using an API Key passed in the `X-API-Key` request header.\n\n## Base URL\n\nAll versioned resources are prefixed with `/api/v2`.",
    "title": "provides the Swagger/OpenAPI specification metadata for the Nexus Tasks API.",
    "version": "2.0.0"
  },
  "host": "localhost:8000",
  "basePath": "/api/v2",
  "paths": {
...
    "/workspaces/{workspaceId}/teams": {
      "get": {
        "description": "Returns all teams that belong to the specified workspace.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "teams"
        ],
        "summary": "List teams in a workspace",
        "operationId": "listTeams",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "description": "UUID of the workspace.",
            "name": "workspaceId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "List of teams.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ResTeam"
              }
            }
          },
          "400": {
            "description": "Invalid workspace ID."
          },
          "401": {
            "description": "Not authenticated."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ]
      },
      "post": {
        "description": "Creates a new team within the specified workspace. Requires workspace admin privileges.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "teams"
        ],
        "summary": "Create a team",
        "operationId": "createTeam",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "description": "UUID of the workspace.",
            "name": "workspaceId",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ReqCreateTeam"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Team created.",
            "schema": {
              "$ref": "#/definitions/ResTeam"
            }
          },
          "400": {
            "description": "Validation error."
          },
          "401": {
            "description": "Not authenticated."
          },
          "403": {
            "description": "Insufficient permissions."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/workspaces/{workspaceId}/teams/{teamId}": {
      "get": {
        "description": "Returns the details of a single team, including its associated projects.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "teams"
        ],
        "summary": "Get a team by ID",
        "operationId": "getTeam",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "description": "UUID of the workspace.",
            "name": "workspaceId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "format": "uuid",
            "description": "UUID of the team.",
            "name": "teamId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Team details with associated projects.",
            "schema": {
              "$ref": "#/definitions/ResTeamWithProjects"
            }
          },
          "400": {
            "description": "Invalid team ID."
          },
          "401": {
            "description": "Not authenticated."
          },
          "404": {
            "description": "Team not found."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ]
      },
      "delete": {
        "description": "Permanently removes a team from the workspace. Requires workspace admin privileges.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "teams"
        ],
        "summary": "Delete a team",
        "operationId": "deleteTeam",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "workspaceId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "format": "uuid",
            "name": "teamId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Team deleted."
          },
          "400": {
            "description": "Invalid team or workspace ID."
          },
          "401": {
            "description": "Not authenticated."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ]
      },
      "patch": {
        "description": "Updates the name and/or description of a team. Requires workspace admin privileges.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "teams"
        ],
        "summary": "Update a team",
        "operationId": "updateTeam",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "workspaceId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "format": "uuid",
            "name": "teamId",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ReqUpdateTeam"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Team updated.",
            "schema": {
              "$ref": "#/definitions/ResTeam"
            }
          },
          "400": {
            "description": "Invalid request."
          },
          "401": {
            "description": "Not authenticated."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/workspaces/{workspaceId}/teams/{teamId}/members": {
      "get": {
        "description": "Returns all members of the specified team along with their roles.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "teams"
        ],
        "summary": "List team members",
        "operationId": "listTeamMembers",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "workspaceId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "format": "uuid",
            "name": "teamId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "List of team members.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ResTeamMember"
              }
            }
          },
          "400": {
            "description": "Invalid team ID."
          },
          "401": {
            "description": "Not authenticated."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ]
      },
      "post": {
        "description": "Adds an existing workspace member to the specified team with an assigned role.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "teams"
        ],
        "summary": "Add a member to a team",
        "operationId": "addTeamMember",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "workspaceId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "format": "uuid",
            "name": "teamId",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ReqAddTeamMember"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Member added successfully."
          },
          "400": {
            "description": "Validation error or invalid path parameter."
          },
          "401": {
            "description": "Not authenticated."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/workspaces/{workspaceId}/teams/{teamId}/members/{uid}": {
      "delete": {
        "description": "Removes the specified user from the team. Their workspace membership is not affected.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "teams"
        ],
        "summary": "Remove a member from a team",
        "operationId": "removeTeamMember",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "workspaceId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "format": "uuid",
            "name": "teamId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "format": "uuid",
            "description": "UUID of the user to remove.",
            "name": "uid",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Member removed."
          },
          "400": {
            "description": "Invalid path parameter."
          },
          "401": {
            "description": "Not authenticated."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ]
      }
    }
  },
  "definitions": {
    "Project": {
      "type": "object",
      "title": "Project represents a project in a workspace.",
      "properties": {
        "created_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "CreatedAt"
        },
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "id": {
          "type": "string",
          "format": "uuid",
          "x-go-name": "ID"
        },
        "is_archived": {
          "type": "boolean",
          "x-go-name": "IsArchived"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "UpdatedAt"
        },
        "workspace_id": {
          "type": "string",
          "format": "uuid",
          "x-go-name": "WorkspaceID"
        }
      },
      "x-go-package": "github.com/AshvinBambhaniya/nexus-tasks/v2/models"
    },
...
    "ReqAddTeamMember": {
      "type": "object",
      "title": "ReqAddTeamMember defines the request payload for adding a member to a team.",
      "required": [
        "email",
        "role"
      ],
      "properties": {
        "email": {
          "description": "Email of the user to add.",
          "type": "string",
          "x-go-name": "Email",
          "example": "dev@example.com"
        },
        "role": {
          "description": "Role to assign to the member. Must be a valid team role enum value.",
          "type": "string",
          "x-go-name": "Role",
          "example": "MEMBER"
        }
      },
      "x-go-package": "github.com/AshvinBambhaniya/nexus-tasks/v2/pkg/structs"
    },
...
    "ReqCreateTeam": {
      "type": "object",
      "title": "ReqCreateTeam defines the request payload for creating a team.",
      "required": [
        "name"
      ],
      "properties": {
        "description": {
          "description": "Optional team description.",
          "type": "string",
          "x-go-name": "Description",
          "example": "Responsible for all server-side services"
        },
        "name": {
          "description": "Team name.",
          "type": "string",
          "x-go-name": "Name",
          "example": "Backend Squad"
        }
      },
      "x-go-package": "github.com/AshvinBambhaniya/nexus-tasks/v2/pkg/structs"
    },
...
    "ReqUpdateTeam": {
      "type": "object",
      "title": "ReqUpdateTeam defines the request payload for updating a team.",
      "properties": {
        "description": {
          "description": "New team description.",
          "type": "string",
          "x-go-name": "Description",
          "example": "Handles platform infrastructure"
        },
        "name": {
          "description": "New team name.",
          "type": "string",
          "x-go-name": "Name",
          "example": "Platform Team"
        }
      },
      "x-go-package": "github.com/AshvinBambhaniya/nexus-tasks/v2/pkg/structs"
    },
...
    "ResTeam": {
      "type": "object",
      "title": "ResTeam defines the response payload for a team.",
      "properties": {
        "description": {
          "description": "Team description.",
          "type": "string",
          "x-go-name": "Description",
          "example": "Responsible for all server-side services"
        },
        "id": {
          "description": "Unique team identifier.",
          "type": "string",
          "format": "uuid",
          "x-go-name": "ID",
          "example": "550e8400-e29b-41d4-a716-446655440000"
        },
        "name": {
          "description": "Team name.",
          "type": "string",
          "x-go-name": "Name",
          "example": "Backend Squad"
        },
        "workspace_id": {
          "description": "UUID of the workspace this team belongs to.",
          "type": "string",
          "format": "uuid",
          "x-go-name": "WorkspaceID",
          "example": "550e8400-e29b-41d4-a716-446655440000"
        }
      },
      "x-go-package": "github.com/AshvinBambhaniya/nexus-tasks/v2/pkg/structs"
    },
    "ResTeamMember": {
      "type": "object",
      "title": "ResTeamMember defines the response payload for a team member.",
      "properties": {
        "role": {
          "$ref": "#/definitions/TeamRole"
        },
        "team_id": {
          "description": "UUID of the team.",
          "type": "string",
          "format": "uuid",
          "x-go-name": "TeamID",
          "example": "550e8400-e29b-41d4-a716-446655440000"
        },
        "user": {
          "$ref": "#/definitions/ResUser"
        },
        "user_id": {
          "description": "UUID of the member user.",
          "type": "string",
          "format": "uuid",
          "x-go-name": "UserID",
          "example": "550e8400-e29b-41d4-a716-446655440001"
        }
      },
      "x-go-package": "github.com/AshvinBambhaniya/nexus-tasks/v2/pkg/structs"
    },
    "ResTeamWithProjects": {
      "type": "object",
      "title": "ResTeamWithProjects defines the response payload for a team with its associated projects.",
      "properties": {
        "description": {
          "description": "Team description.",
          "type": "string",
          "x-go-name": "Description",
          "example": "Responsible for all server-side services"
        },
        "id": {
          "description": "Unique team identifier.",
          "type": "string",
          "format": "uuid",
          "x-go-name": "ID",
          "example": "550e8400-e29b-41d4-a716-446655440000"
        },
        "name": {
          "description": "Team name.",
          "type": "string",
          "x-go-name": "Name",
          "example": "Backend Squad"
        },
        "projects": {
          "description": "Projects associated with this team.",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Project"
          },
          "x-go-name": "Projects"
        },
        "workspace_id": {
          "description": "UUID of the workspace this team belongs to.",
          "type": "string",
          "format": "uuid",
          "x-go-name": "WorkspaceID",
          "example": "550e8400-e29b-41d4-a716-446655440000"
        }
      },
      "x-go-package": "github.com/AshvinBambhaniya/nexus-tasks/v2/pkg/structs"
    },
...
    "ResUser": {
      "type": "object",
      "title": "ResUser defines the response payload for a user.",
      "properties": {
        "email": {
          "description": "User's email address.",
          "type": "string",
          "x-go-name": "Email",
          "example": "user@example.com"
        },
        "full_name": {
          "description": "User's display name.",
          "type": "string",
          "x-go-name": "FullName",
          "example": "Jane Doe"
        },
        "id": {
          "description": "Unique user identifier.",
          "type": "string",
          "format": "uuid",
          "x-go-name": "ID",
          "example": "550e8400-e29b-41d4-a716-446655440000"
        },
        "is_active": {
          "description": "Whether the user account is active.",
          "type": "boolean",
          "x-go-name": "IsActive",
          "example": true
        }
      },
      "x-go-package": "github.com/AshvinBambhaniya/nexus-tasks/v2/pkg/structs"
    },
...
    "TeamRole": {
      "type": "string",
      "title": "TeamRole defines the role of a user in a team.",
      "x-go-package": "github.com/AshvinBambhaniya/nexus-tasks/v2/models"
    },
...
  },
...
  "securityDefinitions": {
    "apiKeyAuth": {
      "description": "A long-lived API key generated via the `/api/v2/auth/api-keys` endpoint. Intended for machine-to-machine access.",
      "type": "apiKey",
      "name": "X-API-Key",
      "in": "header"
    },
    "cookieAuth": {
      "description": "HTTP-only session cookie set by the login/register endpoints. The cookie name is `nexus_user` and contains a signed JWT.\n",
      "type": "apiKey",
      "name": "Cookie",
      "in": "header"
    }
  },
  "security": [
    {
      "cookieAuth": []
    },
    {
      "apiKeyAuth": []
    }
  ]
}